+++ src/App.js
@@ -1,8 +1,8 @@
 import './App.css';
 import { BrowserRouter } from 'react-router-dom';
 
-import { Header } from './components/Header';
-import { Footer } from './components/Footer';
+import { Header } from './components/Header/header';
+import { Footer } from './components/Footer/footer';
 
 import Routes from './Routes';
